--- a/codecs/h264_packet.go
+++ b/codecs/h264_packet.go
@@ -7,39 +7,85 @@ const (
 	fuaHeaderSize = 2
 )
 
-func emitNalus(nals []byte, emit func([]byte)) {
-	nextInd := func(nalu []byte, start int) (indStart int, indLen int) {
-		zeroCount := 0
+// I am sorry world, for the OPTIMIZATIONS I have bestowed upon thee.
+// This parses an Annex.B bitstream and returns the next NALU and any remaining data.
+// Annex.B uses a start-code of either 001 or 0001.
+// Each NALU is prefixed with a start-code, which runs until the next start-code (or EOF)
+func nextNALU(data []byte) (nalu []byte, remaining []byte) {
+	i := 0
 
-		for i, b := range nalu[start:] {
-			if b == 0 {
-				zeroCount++
-				continue
-			} else if b == 1 {
-				if zeroCount >= 2 {
-					return start + i - zeroCount, zeroCount + 1
+	// Loop over the entire data, except for the last 3 bytes.
+	// We know that we can't fit a start code in that size plus it would cause a range error.
+	for i+2 < len(data) {
+		// We want to check if the current index is the start of the start code, otherwise advance.
+
+		// Check the 3rd byte first, because it's the most useful.
+		switch data[i+2] {
+		case 0:
+			// state: ? ? 0
+
+			switch {
+			case data[i+1] != 0:
+				// state: ? n 0
+
+				// The 3rd byte could be the start of a code.
+				i += 2
+			case data[i] != 0:
+				// state: n 0 0
+
+				// The 2nd byte could be the start of a code.
+				i++
+			default:
+				// state: 0 0 0 ?
+
+				// I don't think it's possible to have three zeroes except for a start code.
+				// But to be safe, let's handle edge cases.
+
+				if i+3 < len(data) {
+					switch data[i+3] {
+					case 0:
+						// state: 0 0 0 0
+
+						// This shouldn't be possible but I guess it could be part of a start code.
+						i++
+					case 1:
+						// state: 0 0 0 1
+						return data[:i], data[i+4:]
+					default:
+						// state: 0 0 0 n
+
+						// No start code possible so we can advance 4 bytes.
+						i += 4
+					}
 				}
+
+				// state: 0 0 0 EOF
+
+				return data, nil
 			}
-			zeroCount = 0
+		case 1:
+			// state: ? ? 1
+
+			if data[i] == 0 && data[i+1] == 0 {
+				// state: 0 0 1
+				return data[:i], data[i+3:]
+			}
+
+			// state: n ? 1 or ? n 1
+
+			// Neither of these could be part of a start code.
+			i += 3
+		default:
+			// state: ? ? n
+
+			// If it's not 0 or 1, then there's no start code nor can it be part of a start code.
+			// So most of the time, we advance 3 bytes at once instead of just 1 byte at a time.
+			i += 3
 		}
-		return -1, -1
 	}
 
-	nextIndStart, nextIndLen := nextInd(nals, 0)
-	if nextIndStart == -1 {
-		emit(nals)
-	} else {
-		for nextIndStart != -1 {
-			prevStart := nextIndStart + nextIndLen
-			nextIndStart, nextIndLen = nextInd(nals, prevStart)
-			if nextIndStart != -1 {
-				emit(nals[prevStart:nextIndStart])
-			} else {
-				// Emit until end of stream, no end indicator found
-				emit(nals[prevStart:])
-			}
-		}
-	}
+	// At the end of the file, return what we've got.
+	return data, nil
 }
 
 // Payload fragments a H264 packet across one or more byte arrays
@@ -50,12 +96,20 @@ func (p *H264Payloader) Payload(mtu int, payload []byte) [][]byte {
 		return payloads
 	}
 
-	emitNalus(payload, func(nalu []byte) {
+	for len(payload) > 0 {
+		nalu, remaining := nextNALU(payload)
+		payload = remaining
+
+		if len(nalu) == 0 {
+			// This will only be true for the first start code.
+			continue
+		}
+
 		naluType := nalu[0] & 0x1F
 		naluRefIdc := nalu[0] & 0x60
 
 		if naluType == 9 || naluType == 12 {
-			return
+			continue
 		}
 
 		// Single NALU
@@ -63,7 +117,7 @@ func (p *H264Payloader) Payload(mtu int, payload []byte) [][]byte {
 			out := make([]byte, len(nalu))
 			copy(out, nalu)
 			payloads = append(payloads, out)
-			return
+			continue
 		}
 
 		// FU-A
@@ -87,7 +141,7 @@ func (p *H264Payloader) Payload(mtu int, payload []byte) [][]byte {
 		naluDataRemaining := naluDataLength
 
 		if min(maxFragmentSize, naluDataRemaining) <= 0 {
-			return
+			continue
 		}
 
 		for naluDataRemaining > 0 {
@@ -124,7 +178,7 @@ func (p *H264Payloader) Payload(mtu int, payload []byte) [][]byte {
 			naluDataIndex += currentFragmentSize
 		}
 
-	})
+	}
 
 	return payloads
 }