--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,5 @@
 module github.com/pions/rtp
 
 go 1.12
+
+require github.com/stretchr/testify v1.3.0