It's not a bottleneck, I just thought there was a bug and then got
carried away. It's probably faster on larger NALUs.

As compared to `emitNalu`:
```
name        old time/op  new time/op  delta
nextNALU-8  40.1ns ± 0%  22.6ns ± 1%  -43.48%
```