@Sean-Der 2020-04-12
Mass closing things with over 1 year of inactivity. If anyone is interested this looks like a really fantastic change, probably just need to rebase and should be fine :)